# LPD-61566 Performance test when getting object entries with nested fields object relationship names