--- a/modules/apps/object/object-api/src/main/java/com/liferay/object/service/ObjectEntryLocalServiceUtil.java
+++ b/modules/apps/object/object-api/src/main/java/com/liferay/object/service/ObjectEntryLocalServiceUtil.java
@@ -371,6 +371,16 @@ public class ObjectEntryLocalServiceUtil {
 			start, end);
 	}
 
+	public static List<ObjectEntry> getManyToManyObjectEntries(
+			long groupId, long objectRelationshipId, Long[] primaryKeys,
+			boolean related, boolean reverse, String search, int start, int end)
+		throws PortalException {
+
+		return getService().getManyToManyObjectEntries(
+			groupId, objectRelationshipId, primaryKeys, related, reverse,
+			search, start, end);
+	}
+
 	public static int getManyToManyObjectEntriesCount(
 			long groupId, long objectRelationshipId, long primaryKey,
 			boolean related, boolean reverse, String search)
@@ -378,6 +388,16 @@ public class ObjectEntryLocalServiceUtil {
 
 		return getService().getManyToManyObjectEntriesCount(
 			groupId, objectRelationshipId, primaryKey, related, reverse,
+			search);
+	}
+
+	public static int getManyToManyObjectEntriesCount(
+			long groupId, long objectRelationshipId, Long[] primaryKeys,
+			boolean related, boolean reverse, String search)
+		throws PortalException {
+
+		return getService().getManyToManyObjectEntriesCount(
+			groupId, objectRelationshipId, primaryKeys, related, reverse,
 			search);
 	}
 
@@ -538,6 +558,16 @@ public class ObjectEntryLocalServiceUtil {
 			end);
 	}
 
+	public static List<ObjectEntry> getOneToManyObjectEntries(
+			long groupId, long objectRelationshipId, Long[] primaryKeys,
+			boolean related, String search, int start, int end)
+		throws PortalException {
+
+		return getService().getOneToManyObjectEntries(
+			groupId, objectRelationshipId, primaryKeys, related, search, start,
+			end);
+	}
+
 	public static int getOneToManyObjectEntriesCount(
 			long groupId, long objectRelationshipId, long primaryKey,
 			boolean related, String search)
@@ -545,6 +575,15 @@ public class ObjectEntryLocalServiceUtil {
 
 		return getService().getOneToManyObjectEntriesCount(
 			groupId, objectRelationshipId, primaryKey, related, search);
+	}
+
+	public static int getOneToManyObjectEntriesCount(
+			long groupId, long objectRelationshipId, Long[] primaryKeys,
+			boolean related, String search)
+		throws PortalException {
+
+		return getService().getOneToManyObjectEntriesCount(
+			groupId, objectRelationshipId, primaryKeys, related, search);
 	}
 
 	public static ObjectEntry getOrAddEmptyObjectEntry(