--- a/modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryLocalServiceImpl.java
+++ b/modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryLocalServiceImpl.java
@@ -4022,13 +4022,19 @@ public class ObjectEntryLocalServiceImpl
 		ObjectField objectField = _objectFieldPersistence.fetchByPrimaryKey(
 			objectRelationship.getObjectFieldId2());
 
-		ObjectEntry objectEntry = fetchObjectEntry(primaryKey);
+		ObjectEntry objectEntry = null;
+
+		if (primaryKeys.length == 1) {
+			objectEntry = fetchObjectEntry(primaryKeys[0]);
+		}
 
 		DynamicObjectDefinitionTable rootDynamicObjectDefinitionTable =
 			_getRootDynamicObjectDefinitionTable(objectEntry);
 
 		Column<DynamicObjectDefinitionTable, Long> primaryKeyColumn =
 			dynamicObjectDefinitionTable.getPrimaryKeyColumn();
+
+		ObjectEntry finalObjectEntry = objectEntry;
 
 		return fromStep.from(
 			dynamicObjectDefinitionTable
@@ -4101,11 +4107,11 @@ public class ObjectEntryLocalServiceImpl
 
 					long rootObjectDefinitionId = 0L;
 
-					if ((objectEntry != null) &&
-						(objectEntry.getRootObjectEntryId() != 0)) {
+					if ((finalObjectEntry != null) &&
+						(finalObjectEntry.getRootObjectEntryId() != 0)) {
 
 						ObjectEntry rootObjectEntry = fetchObjectEntry(
-							objectEntry.getRootObjectEntryId());
+							finalObjectEntry.getRootObjectEntryId());
 
 						rootObjectDefinitionId =
 							rootObjectEntry.getObjectDefinitionId();