--- a/modules/apps/object/object-test/src/testIntegration/resources/com/liferay/object/service/test/dependencies/object-entry-performance.properties
+++ b/modules/apps/object/object-test/src/testIntegration/resources/com/liferay/object/service/test/dependencies/object-entry-performance.properties
@@ -2,3 +2,4 @@ object.entries.count=500
 object.entries.delete.max.time=30000
 object.entries.get.max.time=60000
 object.entries.import.max.time=36000
+object.relationships.count=3