Performance test when a non-admin user is getting object entries with the nested field of object relationship names.

| Run #   | Without Solution (ms) | With Solution (ms) |
|---------|-----------------------|--------------------|
| Run 1  | 1492                  | 456                |
| Run 2  | 1136                  | 354                |
| Run 3  | 1206                  | 341                |
| Run 4  | 1057                  | 317                |
| Run 5  | 1108                  | 348                |
| Run 6  | 1053                  | 334                |
| Run 7  | 1118                  | 295                |
| Run 8  | 1047                  | 331                |
| Run 9  | 1036                  | 294                |
| Run 10 | 1007                  | 306                |