@liferay-continuous-integration 2025-08-13
<html><h3>:heavy_check_mark: ci:test:sf - 1 out of 1 jobs passed in 3 minutes</h3><details><summary>Click here for more details.</summary><h4>Base Branch:</h4><p>Branch Name: <a href="https://github.com/liferay/liferay-portal/tree/master">master</a><br/>Branch GIT ID: <a href="https://github.com/liferay/liferay-portal/commit/9c99eed392f5116d68321461d7cffd6507a9f8a8">9c99eed392f5116d68321461d7cffd6507a9f8a8</a></p><h4>Sender Branch:</h4><p>Branch Name: <a href="https://github.com/marcelabc/liferay-portal/tree/LPD-61566-1">LPD-61566-1</a><br/>Branch GIT ID: <a href="https://github.com/marcelabc/liferay-portal/commit/2160b11c14b54520fb0eeb4d6676654fe952cbe2">2160b11c14b54520fb0eeb4d6676654fe952cbe2</a></p>1 out of 1jobs PASSED<details><summary><strong>1 Successful Jobs:</strong></summary><ul><li><a href="https://test-1-31.liferay.com/job/test-portal-source-format/7400/">test-portal-source-format</a></li></ul></details><h5>For more details click <a href="https://test-1-31.liferay.com/userContent/jobs/test-portal-source-format/builds/7400/jenkins-report.html">here</a>.</h5></details></html>