@liferay-continuous-integration 2025-08-13
<div><div><strong>Jenkins Build:</strong><span></span><a href="https://test-1-31.liferay.com/job/test-portal-source-format/7400">test-portal-source-format#7400</a><br/></div><div><strong>Jenkins Report:</strong><span></span><a href="https://test-1-31.liferay.com/userContent/jobs/test-portal-source-format/builds/7400/jenkins-report.html">jenkins-report.html</a><br/></div><div><strong>Jenkins Suite:</strong><span></span>sf<br/></div><div><strong>Pull Request:</strong><span></span><a href="https://github.com/liferay-bpm/liferay-portal/pull/4966">liferay-bpm#4966</a><br/></div><div><strong>Testray Routine:</strong><span></span><a href="https://testray.liferay.com/#/project/35392/routines/45357">EE Pull Request</a><br/></div><div><strong>Testray Build:</strong><span></span><a href="https://testray.liferay.com/#/project/35392/routines/45357/build/288668932">[master] ci:test:sf - marcelabc &gt; liferay-bpm - PR#4966 - 2025-08-13[10:10:38]</a><br/></div><div><strong>Testray Build ID:</strong><span></span><code>288668932</code><br/></div><div><strong>Testray Importer:</strong><span></span><a href="https://test-1-31.liferay.com/job/test-portal-source-format/7400/">test-portal-source-format#7400</a><br/></div></div>